pyproject.toml
[project]
name = "fast_api"
version = "0.1.0"
requires-python = ">=3.12,<4.0"
description = "WeatherVault — weather app with auth and saved history"
authors = []
readme = "README.md"
dependencies = [
  "fastapi>=0.122.0,<1",
  "uvicorn[standard]>=0.38.0,<1",
  "pydantic>=2.12.5,<3",
  "pydantic[email]>=2.12.5,<3",
  "pydantic-settings>=2.12.0,<3",
  "yarl>=1.22.0,<2",
  "ujson>=5.11.0,<6",
  "SQLAlchemy[asyncio]>=2.0.44,<3",
  "alembic>=1.17.2,<2",
  "asyncpg>=0.31.0,<1",
  "httptools>=0.7.1,<1",
  "python-jose[cryptography]>=3.3.0,<4",
  "passlib[bcrypt]>=1.7.4,<2",
  "bcrypt==4.0.1",
  "httpx>=0.27.0,<1",
]

[dependency-groups]
dev = [
  "pytest>=9.0.1,<10",
  "ruff>=0.14.6,<1",
  "mypy>=1.19.0,<2",
  "pre-commit>=4.5.0,<5",
  "pytest-cov>=7.0.0,<8",
  "anyio>=4.11.0,<5",
  "pytest-env>=1.2.0,<2",
  "httpx>=0.28.1,<1",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
addopts = "-p no:unraisableexception"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
anyio_mode = "auto"
env = [
    "FAST_API_ENVIRONMENT=pytest",
    "FAST_API_DB_BASE=fast_api_test",
]

[tool.ruff]
target-version = "py313"
lint.select = ["E", "F", "W", "I", "N", "D", "ANN", "S", "B", "C4", "PIE", "T20", "Q", "RSE", "RET", "SIM", "PTH", "PL", "RUF", "UP"]
lint.ignore = ["D105", "D107", "B008", "D211", "D212", "D401", "D104", "D100", "D202", "ANN401", "PLR0913", "D106"]
exclude = ["fast_api/db/migrations", ".venv/"]
line-length = 88
